day2/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
class-registry==2.1.2
cryptography==2.6.1
filters==1.3.2
idna==2.8
paho-mqtt==1.4.0
pycparser==2.19
pyOpenSSL==19.0.0
PyOTA==2.0.7
pyserial==3.4
pysha3==1.0.2
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.1
regex==2019.3.12
requests==2.21.0
six==1.12.0
urllib3==1.24.1
zmq==0.0.0
